buffer RX input reports
When we call read_timeout to fill the report buffer,
if we specify a size smaller than the max input report length,
the report can become truncated. The report will indicate that
it has length M, where the read size is N; when M > N,
the data would be lost.

So we need to buffer up to the max input report length to avoid
losing data.

This also fixes a bug that resulted in a panic: the existing code
might try to fill the last N bytes of the data output buffer,
while in actuality attempting to write M bytes to the data slice,
resulting in an out-of-range panic.